requirements.txt
openai>=1.12.0
streamlit>=1.28.0
faiss-cpu>=1.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
pypdf>=4.0.0
